[sync]refactor test.
